.history/login-service/requirements_20250120030059.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
pydantic
python-multipart
pydantic[email]

